Вынес фильтр comments по id в api.remove, изменил отображение аватарки, вынес validatorConfig в папку с компонентом, изменил isLoding, удалил проверку optionsArray
