+++ src/app/components/UI/userAvatar.jsx
@@ -1,19 +1,20 @@
 import React from "react";
+import PropTypes from "prop-types";
 
-const UserAvatar = () => {
+const UserAvatar = ({ user }) => {
     return (
         <img
-            src={`https://avatars.dicebear.com/api/avataaars/${(
-                Math.random() + 1
-            )
-                .toString(36)
-                .substring(7)}.svg`}
+            src={`https://avatars.dicebear.com/api/avataaars/${user?.name}.svg`}
             className="rounded-circle shadow-1-strong me-3"
             alt="avatar"
             width="65"
             height="65"
         />
     );
+};
+
+UserAvatar.propTypes = {
+    user: PropTypes.object
 };
 
 export default UserAvatar;
